+++ apis/python/src/tiledbsoma/__init__.py
@@ -157,7 +157,7 @@ from ._general_utilities import (
     get_storage_engine,
     show_package_versions,
 )
-from ._index_util import tiledbsoma_build_index
+from ._indexer import tiledbsoma_build_index
 from ._measurement import Measurement
 from ._sparse_nd_array import SparseNDArray
 from .options import SOMATileDBContext, TileDBCreateOptions
