+++ apis/python/src/tiledbsoma/_experiment.py
@@ -13,7 +13,7 @@ from typing_extensions import Self
 
 from ._collection import Collection, CollectionBase
 from ._dataframe import DataFrame
-from ._index_util import tiledbsoma_build_index
+from ._indexer import tiledbsoma_build_index
 from ._measurement import Measurement
 from ._tdb_handles import Wrapper
 from ._tiledb_object import AnyTileDBObject
