+++ apis/python/tests/test_indexer.py
@@ -6,7 +6,7 @@ import pandas as pd
 import pyarrow as pa
 import pytest
 
-from tiledbsoma._index_util import tiledbsoma_build_index
+from tiledbsoma._indexer import tiledbsoma_build_index
 from tiledbsoma.options import SOMATileDBContext
 from tiledbsoma.options._soma_tiledb_context import _validate_soma_tiledb_context
